:recycle: Refactor form serializers file
