+++ src/openforms/forms/api/serializers/form_version.py
@@ -1,0 +1,18 @@
+from rest_framework import serializers
+
+from openforms.accounts.api.serializers import UserSerializer
+
+from ...models import FormVersion
+
+
+class FormVersionSerializer(serializers.HyperlinkedModelSerializer):
+    user = UserSerializer(required=False, read_only=True)
+
+    class Meta:
+        model = FormVersion
+        fields = (
+            "uuid",
+            "created",
+            "user",
+            "description",
+        )
